Update OG's personal page
