+++ .gitignore
@@ -4,3 +4,4 @@
 .Ruserdata
 
 /.quarto/
+.DS_Store
