+++ posts/members/OG/index.qmd
@@ -22,7 +22,7 @@ format:
 ---
 
 <div style="text-align: center;">
-  ![](../../../images/OG.jpg)
+  <img src="../../../images/OG.jpg" style="width: 50%;" />
 </div>
 
 ## Biography
@@ -36,8 +36,9 @@ Olivier Gourgue is a marine ecosystem modeler, specialized in the development of
 - **Boston University** | Boston, MA, USA<br />
   Visiting Scholar | 2019-2021
 - **Vrije Universiteit Brusssel** | Brussels, Belgium<br />
-  **Flanders Hydraulics** | Antwerp, Belgium<br />
   Postdoctoral Researcher | 2013-2016
+- **Flanders Hydraulics** | Antwerp, Belgium<br />
+  Consultant | 2013-2016
 - **UCLouvain** | Louvain-la-Neuve, Belgium<br />
   Postdoctoral Researcher | 20011-2013<br />
   Teaching Assistant | 2006-2012<br />
@@ -50,13 +51,13 @@ Olivier Gourgue is a marine ecosystem modeler, specialized in the development of
   BSc Engineering Science | 2000-2002
 
 ## Projects
-- [COHERENS](../../projects/COHERENS): COupled Hydrodynamical-Ecological model for REgioNal and Shelf seas.
+- [COHERENS:](../../projects/COHERENS) COupled Hydrodynamical-Ecological model for REgioNal and Shelf seas.
 
 ## Publications
 ### 2025
 - L. Alaerts, J. Lambrechts, N.R. Randresihaja, L. Vandenbulcke, **O. Gourgue**, E. Hanert, M. Grégoire, An integrated high-resolution bathymetric model for the Danube Delta system, *Earth System Science Data Discussion* (accepted). [DOI](https://doi.org/10.5194/essd-2024-529)
 - D. Wang, C. Gu, S. Temmerman, J.-P. Belliard, **O. Gourgue**, L. Xue, J. Bai, Coastal marsh vulnerability to sea-level rise is exacerbated by plant species invasion, *Global Change Biology*, 31, 2, e70058. [DOI](https://doi.org/10.1111/gcb.70058)
-- D. Wang, J. Bai, C. Gu, **O. Gourgue**, J.-P. Belliard, L. Cui, Y. Ke, L. Xue, L. Wen, S. Temmerman, How does landscape vegetation configuration regulate local channel initiation in a rapidly expanding micro-tidal system? *Journal of Hydrology*, 649, 132473. [DOI](https://doi.org/10.1016/j.jhydrol.2024.132473)
+- D. Wang, J. Bai, C. Gu, **O. Gourgue**, J.-P. Belliard, L. Cui, Y. Ke, L. Xue, L. Wen, S. Temmerman, How does landscape vegetation configuration regulate local channel initiation in a rapidly expanding micro-tidal system? *Journal of Hydrology*, 649, 132473. [DOI](https://doi.org/10.1016/j.jhydrol.2024.132473) 
 
 ### 2024
 - **O. Gourgue**, J.-P. Belliard, Y. Xu, M.G. Kleinhans, S. Fagherazzi, S. Temmerman, Dense vegetation hinders sediment transport towards saltmarsh interiors, *Limnology and Oceanography Letters*, 9, 6, 764-775. [DOI](https://doi.org/10.1002/lol2.10436)
@@ -110,3 +111,8 @@ Modelling and Software*, 24, 8, 969-981. [DOI](http://doi.org/10.1016/j.envsoft.
 
 ### 2007
 - **O. Gourgue**, E. Deleersnijder, L. White, Toward a generic method for studying water renewal, with application to the epilimnion of Lake Tanganyika, *Estuarine, Coastal and Shelf Science*, 74, 4, 628-640. [DOI](http://doi.org/10.1016/j.ecss.2007.05.009)
+
+## Code and software
+- [Demeter:](https://github.com/ogourgue/demeter) Multiscale biogeomorphic modeling framework in Python.
+- [TidalGeoPro:](https://github.com/ogourgue/tidalgeopro) Python package to analyze geometric and hydraulic properties of tidal channels.
+- [VFC:](https://github.com/ogourgue/vfc) Vegetation/Flow Convolution Python module.
